backend/requirements.txt
alphashape==1.3.1
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
comm==0.1.3
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
exceptiongroup==1.1.1
executing==1.2.0
face-alignment==1.3.5
filelock==3.10.2
Flask==2.2.3
Flask-PyMongo==2.3.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.25.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
jwt==1.3.1
kiwisolver==1.4.4
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.0
numba==0.56.4
numpy==1.23.5
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pymongo==4.3.3
pyngrok==5.2.1
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
Rtree==1.0.1
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
sklearn==0.0
smmap==5.0.0
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2023.2.3
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trimesh==3.20.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.14.0
